Cleaned up Rope viable platforms in Demo3 Tutoriallevel
When quiried without a value having been set, the volume functions will now return a default of 0.5, not 0.
The dictionary of corners the rope wraps around is now reset on KillPlayer.
Demo3 Tutoriallevel now moves to Demo3 level1 on completion
Fixed error where the pretty rope did not form sometimes.